requirements.txt
torch
torchvision
numpy
matplotlib
opencv-python
Pillow
scikit-learn
python-docx
